+++ app/src/main/java/woowacourse/paint/model/brush/Brush.kt
@@ -1,0 +1,14 @@
+package woowacourse.paint.model.brush
+
+import android.graphics.Color
+import android.graphics.Paint
+
+sealed class Brush : BrushSetting {
+    val paint get() = paintInstance
+
+    companion object {
+        val paintInstance = Paint().apply {
+            color = Color.RED
+        }
+    }
+}
