+++ app/src/main/java/woowacourse/paint/model/brush/BrushSetting.kt
@@ -1,0 +1,5 @@
+package woowacourse.paint.model.brush
+
+interface BrushSetting {
+    fun setStyle()
+}
