+++ app/src/main/java/woowacourse/paint/model/brush/Eraser.kt
@@ -1,0 +1,5 @@
+package woowacourse.paint.model.brush
+
+object Eraser : Brush() {
+    override fun setStyle() = Unit
+}
