+++ app/src/main/java/woowacourse/paint/model/brush/Pen.kt
@@ -1,0 +1,13 @@
+package woowacourse.paint.model.brush
+
+import android.graphics.Paint
+
+object Pen : Brush() {
+    override fun setStyle() {
+        paint.apply {
+            style = Paint.Style.STROKE
+            strokeCap = Paint.Cap.ROUND
+            strokeJoin = Paint.Join.ROUND
+        }
+    }
+}
