+++ app/src/main/java/woowacourse/paint/model/brush/Rectangle.kt
@@ -1,0 +1,9 @@
+package woowacourse.paint.model.brush
+
+import android.graphics.Paint
+
+object Rectangle : Brush() {
+    override fun setStyle() {
+        paint.style = Paint.Style.FILL
+    }
+}
